Add supply kontainers in KIS SC-62 form factor
These two new parts are based on the existing RadialSupPack and
RadialLqdTank, morphed into the form factor of the SC-62 storage
container from KIS.  Like the SC-62, they can be carried on a kerbal's
back and mounted (even by non-engineers) on the SM-62 container mount.
Capacity, mass, and cost have been scaled down to 80% of the
RadialSupPack and RadialLqdTank, to match the SC-62's 1000L volume.

For now, these parts use the same model as the SC-62, but they probably
ought to have models of their own before they're considered "finished".