Add models for mountable kontainers
This adds three new models:

 * MountableBase.mu: Basically the bottom portion of the KIS SC-62
   container, which is shaped to fit on the SM-62 mounting bracket.
 * MountableLqdTank.mu: A group of four capsule-shaped tanks in the same
   style as the MKV_SaddleTank model.
 * MountableSupPack.mu: A single storage crate from the Karibou's
   KER_Crate model.

The parts' .cfg files each have two MODEL blocks, placing the
appropriate kontainer model on top of the common mounting base.

I've added a copy of RoverKontainers.dds from Karibou, since it's needed
by the MountableSupPack model.  For the MountableBase, I added a pair of
dummy single-pixel placeholders for KIS textures referenced by the
model; the .cfg file has "texture" attributes that replace those with
the real textures from the user's KIS installation (which is safe since
these parts depend on KIS).  The MountableLqdTank uses existing
Kontainers textures already present in the Assets folder.

Note that the origin in each .mu file is at the bottom of the part, and
there are "position" attributes in the .cfg file to translate them
downward so the origin is within the part instead.  This is so that the
part's center of mass is defined in the .cfg file instead of the .mu.
The idea is that there might be other parts in the future that use the
same MountableBase model with a taller or shorter object on top, so
MountableBase.mu shouldn't have any built-in assumption about where the
part's center of mass is.