Adjust collider to avoid clipping SM-62
A recent KIS update changed the SC-62 container's collider so it doesn't
intersect that of the SM-62 container mount, so that the two parts don't
push each other apart when decoupled using the SM-62's "Release" button.
This change makes the same adjustment to the MountableBase collider, for
the same reason.