Fix part positions to match KIS container
The KIS_Container1 model has its origin 0.247985 units from the bottom
of the part.  For the mountable kontainer models, I'd placed the origin
at the bottom since it makes the vertex coordinates nicer, then moved it
in-game using a "position" attribute in the config file.  I put it 0.34
units from the bottom because that looked like roughly where the part's
center of mass ought to be, but I hadn't noticed that this difference
affects how the part is positioned when worn on a kerbal's back.  I also
hadn't noticed that the KIS container's attach node isn't exactly at the
bottom of the model; it's slightly above.

To avoid discrepancies in positioning, I've changed the "position" and
"node_stack_bottom" attributes to place the mountable kontainers' origin
and attachment node in the same places as on the KIS container.