New models for mountable kontainers
The old models for these parts were designed to re-use existing textures
so I wouldn't have to make new ones, but they looked pretty awkward as a
result.  It was also awkward that I had to include dummy placeholders
for KIS textures and a copy of a Karibou texture.  These new models have
proper textures of their own, and they look much nicer.