Move large nuke fuel tanks to correct CTT node
In the stock tech tree, the nuclear fuel/waste tanks live in the
"Heavier Rocketry" node.  With the Community Tech Tree installed, the
small tanks live in "Nuclear Fuel Systems" instead, but the newer large
tanks were left behind in "Heavier Rocketry", presumably by accident.

I've updated the MM patch with wildcards on the part names so that it
matches the large variants as well as, hopefully, any other variants
that may be added in the future.