Fix C3_LqdTrussTank initial xenon/argon resource amounts
All the other spherical tanks default to being filled completely when
configured for xenon or argon, but this one was filled only partially.  The
initialResourceAmounts numbers were the same as C3_RTank_01, where they
represent the tank's full capacity, so it looks like the line was probably
copied from that part but not updated to reflect this one having four spheres
instead of one.