Add ModuleCryoTank to liquid kontainers that lacked it
When CryoTanks is installed, all tanks that can hold LqdHydrogen should have a
ModuleCryoTank that makes the hydrogen boil off unless the tank is actively
cooled.  The CryoTanks mod automatically patches all tanks that don't already
have fuel switchers, but it doesn't touch the USI tanks because they have
FSfuelSwitch modules.  The CryoTanks.cfg file here is what adds boiloff to the
USI tanks, but a few tanks had been omitted.  I've added them to the patch so
that everything gets LH2 boiloff as it should.