Reduce MountableSupPack/MountableLqdTank volume to match KIS
KIS updates changed the volume of the SC-62 inventory container from
1000L to 550L (which is more plausible).  The SK-62 and SK-62L kontainer
parts are the same size and shape, so I've changed their volume to
match.

In addition, KIS used to force the SC-62's *external* volume to 1100L as
a kluge to prevent players from putting one instance inside another.
The new internal volume of 550L is a little smaller than the external
volume that KIS calculates automatically from the model's bounding box,
so that workaround was no longer needed and KIS removed it.  Likewise,
I've done the same here.

Mass and cost have been reduced to match the new ratios with
RadialSupPack and RadialLqdTank (now 44% instead of 80%).  Note that KIS
changed the SC-62's mass from 100kg to 10kg, but the SK-62 and SK-62L
are now 68.75kg (down from 125kg).  This is an intentional small
inconsistency with KIS, in favor of consistency with other USI parts.
It seems reasonable to say that resource kontainers are made of sturdier
and heavier stuff than a hiking backpack (which is what the SC-62's part
description says it's based on).