Fix MountableSupPack/MountableLqdTank not appearing on female backpack
Like the SC-62 part from KIS, the SK-62 and SK-62L are meant to appear
like a backpack when carried in a kerbal's inventory, but the old way of
doing that doesn't work on the female kerbal model.  KIS bug #269 fixed
that for the SC-62 and other KIS parts, and I've made the same change
here to the USI parts.